Initial draft of a tool for collecting NOTICE for Apache licensed projects
At this stage, sbom is pulled from GitHub API of the top package.
The list of depedencies is iterated from the sbom to get their
license and notice files in case the license is APACHE-2.0.
The go-packages package is used to obtain their source location
when available. NOTICES are pulled from GitHub or GitLab depending
on provided url. Redirects are handled properly and custom overrides
are included for cases where the main logic fails and we know the
answer.
Huge refactor is required once we better understand the solution, but
mostly the big patterns we start seeing is that exception lists and
override mappings can be extracted to files to simplify significantly
the codebase. Secondly all the pulling functions from different sources
could be mostly extracted into a single one with a few parameters.
And, as we add languages a abstraction of the package manager is
required to reduce repetition.
Test is unexistent so far, if we move to productionize this, mocking
external services will be necessary, pytest offers great tools for that.
Since we are still exploring the APIs we consume we don't have yet good
examples for the mocks results.