Extracted FS operations into mockable functions
Replacing mocked functions as os.listdir and os.walk by functions
wrapping them in the package we own to be able to mock them without
mocking the real os functions.
This is a temporary change, after refactoring our classes in the
near future, those functions are not goign to be required anymore.